+++ libs/LLVM.h
@@ -22,20 +22,26 @@
 #include <llvm/IR/IRBuilder.h>
 #include <llvm/IR/NoFolder.h>
 
+#include <llvm/Analysis/LoopAnalysisManager.h>
+#include <llvm/Analysis/CGSCCPassManager.h>
+#include <llvm/Analysis/CGSCCPassManager.h>
+
+#include <llvm/Passes/PassBuilder.h>
+
 #include <llvm/Transforms/Utils/Cloning.h>
 
 #include <llvm/Support/TargetSelect.h>
 
 
 #include <llvm/IR/LegacyPassManager.h>
+#include <llvm/IR/PassManager.h>
 
 #include <llvm/MC/TargetRegistry.h>
 #include <llvm/Support/FileSystem.h>
-// #include <llvm/Support/TargetSelect.h>
-// #include <llvm/Support/raw_ostream.h>
-// #include <llvm/Target/TargetMachine.h>
-// #include <llvm/Target/TargetOptions.h>
+
 #include <llvm/TargetParser/Host.h>
+
+
 
 #undef _SILENCE_CXX20_CISO646_REMOVED_WARNING